backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic-settings==2.10.1
httpx==0.28.1
pandas==2.3.2
rapidfuzz==3.14.1
google-genai==1.29.0
pytest==8.4.2
